+++ CHANGELOG.md
@@ -1,15 +1,33 @@
-# Ver
+# Version History
 
+## 0.0.4
+
+* adjustment WebviewWrapperController api
+* update webview_flutter: ^4.13.0
+* update 0.0.4
+* fix:
+* on inject object ready callback
+* on onPageStartScriptReady callback
+* on onPageEndScriptReady callback
+
+--- 
 
 ## 0.0.3
+
 * update WebviewWrapperController，
 * add setNavigationDelegateWrapper replace setNavigationDelegate
 * add InjectJsObject injectJsScript,run injectJsScript when inject object
 
+---
+
 ## 0.0.2
-* remove  NavigationDelegate onSslAuthError on webview_flutter 4.12.0
+
+* remove NavigationDelegate onSslAuthError on webview_flutter 4.12.0
 * update readme
 
+---
+
 ## 0.0.1
+
 * init webview wrapper
 * add WebviewWrapperController.addInjectJsObject